virt: increase virtual machine boot delay
On slower machines, the virtual machine may not boot up within 2 seconds.
This allows the boot command to go through even on such machines. Note that
this is a stopgap -- the correct fix should be something like detecting the
boot prompt and only sending the boot command when the machine is ready.