glass: allow local rsync uploads
In some cases, the "upload" process involves placing the binaries in a
given directory instead of SSH-ing into a server, such as in CI where
the binaries have to be transferred out of an isolated environment.
This change allows this by enabling local rsync when the user is not
specified in the upload configuration.